# Server/api-spec/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "0.1.0",
        "title": "KFCSmartCorp",
        "description": "HRM System for SmartDEV"
    },
    "host": "localhost:3007",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/leave/configure/workWeek": {
            "post": {
                "tags": [
                    "Work Week"
                ],
                "summary": "Update work week",
                "description": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "work week",
                        "in": "body",
                        "description": "work week object to add or update",
                        "schema": {
                            "$ref": "#/definitions/works"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "If work country not exits, add object, else update workweek"
                    }
                }
            }
        },
        "/leave/configure/workWeek/{workCountry}": {
            "get": {
                "tags": [
                    "Work Week"
                ],
                "summary": "Get work week arcording country",
                "description": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "workCountry",
                        "description": "The country that needs to be get work week",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "If have error return error notice, else return success notice"
                    }
                }
            }
        },
        "/leave/configure/leaveType": {
            "post": {
                "tags": [
                    "Leave Type"
                ],
                "summary": "Add new leave type",
                "description": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "leave type",
                        "in": "body",
                        "description": "leave type object to add",
                        "schema": {
                            "$ref": "#/definitions/leaveType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "If leave type name or color was existed or have error in save process return error message, else return success message"
                    }
                }
            },
            "get": {
                "tags": [
                    "Leave Type"
                ],
                "summary": "Get all leave type",
                "description": "",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "If have error return error notice, else return success notice"
                    }
                }
            },
            "put": {
                "tags": [
                    "Leave Type"
                ],
                "summary": "Update an leave type",
                "description": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "LeaveType",
                        "in": "body",
                        "description": "LeaveType object to update",
                        "schema": {
                            "$ref": "#/definitions/leaveType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return message error if color existed or meet problem in find or save process, else return update success notice"
                    }
                }
            }
        },
        "/leave/configure/leaveType/findByIdentify/{identify}": {
            "get": {
                "tags": [
                    "Leave Type"
                ],
                "summary": "Get a leave type arcording id",
                "description": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "identify",
                        "description": "The identify to determine leave type",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "If have error return error notice, else return leave type object"
                    }
                }
            }
        },
        "/leave/configure/leaveType/{id}": {
            "delete": {
                "tags": [
                    "Leave Type"
                ],
                "summary": "Delete an leave type",
                "description": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "The id that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "If have error in process return error notice, else return success notice"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "works": {
            "required": [
                "workCountry",
                "workWeeks"
            ],
            "properties": {
                "workCountry": {
                    "type": "string"
                },
                "workWeeks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/workweeks"
                    }
                }
            }
        },
        "workweeks": {
            "required": [
                "monday",
                "tuesday",
                "wednesday",
                "thurday",
                "friday",
                "saturday",
                "sunday"
            ],
            "properties": {
                "monday": {
                    "type": "string"
                },
                "tuesday": {
                    "type": "string"
                },
                "wednesday": {
                    "type": "string"
                },
                "thurday": {
                    "type": "string"
                },
                "friday": {
                    "type": "string"
                },
                "saturday": {
                    "type": "string"
                },
                "sunday": {
                    "type": "string"
                }
            }
        },
        "leaveType": {
            "required": [
                "leavetype",
                "country",
                "_color"
            ],
            "properties": {
                "leavetype": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "_color": {
                    "type": "string"
                }
            }
        },
...
    }
}